int/commands: moves in from `cmd/gonfique` to prepare moving entry point to the module root.
