+++ internal/commands/generate/main.go
